+++ src/os/cpu_freq/mod.rs
@@ -13,14 +13,14 @@ pub use crate::os::cpu_freq::openbsd::*;
 /// The errors that can occur when reading the frequency of the CPU.
 pub enum CpuFreqError {
     CpuNotFound,
-    SysctlCompatError,
+    SysctlCompat,
 }
 
 impl fmt::Display for CpuFreqError {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         match &self {
             CpuFreqError::CpuNotFound => write!(f, "CPU not found"),
-            CpuFreqError::SysctlCompatError => write!(f, "Sysctl compat. error"),
+            CpuFreqError::SysctlCompat => write!(f, "Sysctl compat. error"),
         }
     }
 }
@@ -30,7 +30,7 @@ impl From<SysctlError> for CpuFreqError {
     fn from(value: SysctlError) -> Self {
         match value {
             SysctlError::NotFound => Self::CpuNotFound,
-            _ => Self::SysctlCompatError,
+            _ => Self::SysctlCompat,
         }
     }
 }